+++ client/src/containers/FundingSpaceTypes/FundingSpaceTypes.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { Table, Column } from '@ctoec/component-library';
 import { BackButton } from '../../components/BackButton';
-import { FundingSpaceType, FUNDING_SPACE_TYPES } from './types';
+import { FundingSpaceType } from '../../shared/models';
+import { FUNDING_SPACE_TYPES } from '../../shared/constants';
 
 const FundingSpaceTypes: React.FC = () => {
   const columns: Column<FundingSpaceType>[] = [
